+++ apps/router-demo/router-host-vue3-2100/CHANGELOG.md
@@ -1,5 +1,12 @@
 # host-vue3
 
+## 1.0.18
+
+### Patch Changes
+
+- @module-federation/enhanced@0.5.1
+- @module-federation/bridge-vue3@0.5.1
+
 ## 1.0.17
 
 ### Patch Changes