+++ apps/runtime-demo/3008-runtime-remote/CHANGELOG.md
@@ -1,5 +1,11 @@
 # 3008-runtime-remote
 
+## 1.0.28
+
+### Patch Changes
+
+- @module-federation/enhanced@0.5.1
+
 ## 1.0.27
 
 ### Patch Changes