+++ packages/bridge/bridge-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @module-federation/bridge-react
 
+## 0.5.1
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.5.1
+
 ## 0.5.0
 
 ### Patch Changes